Defined three stacks: token, operation, and data
As I started planning the initial set of words, I realized it would be
easier to center everything around stacks within the VM. A "parse area"
an similar "areas" are yet to be defined.

Current plan is for the scanner to scan input text for tokens, placing
them on the token stack. The parser will pop tokens off the token stack
and translate them into operations puushed to the operation stack.
Interpretation happens by popping operations off of the operations
stack.