Replace use of Vec (stack) with VecDeque (ring buffer)
This commit renames and retypes the Token and Operations collections.
Tokens and Operations have an order, which leads to a pattern of pulling
from the front of their collections or appending to the back of them.
A ring buffer is much better suited for this usage pattern than a stack.

In contrast, operations on the "data stack" normally operate at the top
of the stack by pushing, popping, or modifying upper values. Rarely are
those operations performed on data at or near the bottom of the stack.