cargo: drop preference for split-debuginfo -> off
Tools like LLDB and the backtrace functionality need debug info to be
split from the binary, on macOS. I had assumed this option being off
kept debug information within the binary, as opposed to unpacking it
into a separate file (dSYM) that would be more confusing. Demonstrably,
LLDB does not see the Rust compilation units or other line info and so
breakpoints have not worked with that setting. The setting might still
work to keep debug info in the binary alone, but demonstrably leaving
this option as the default ("packed") fixes issues with debugging.

More research and testing is needed.