repl: low-verbosity mode
Statements are entered which produces a new line. Parsing is silent,
although execution may produce text on the new line. If it does, another
new line is produced. If execution fails, an error is printed on the new
line. If there items on the data stack, the top most value is printed;
if execution succeeded and there is nothing on the stack, the word "ok"
is printed.